+++ RELEASE_NOTES.rst
@@ -36,6 +36,7 @@ Vagga 0.4.0
 * Mount `/dev/shm` by default (needed for ubuntu host, fixes #32)
 * Implement forwarding proxy variables by default (#38)
 * Run ``!Sh`` scripts with ``-ex`` options (#72)
+* Implement ``subdirs`` key for ``!Tmpfs`` volume
 
 .. _kill-unresponsive-after: http://vagga.readthedocs.org/en/latest/commands.html#opt-kill-unresponsive-after